fixing errors from Itests
