+++ modules/engines/engine-hdb/src/main/java/com/sap/xsk/hdb/ds/parser/hdbtable/XSKTableParser.java
@@ -24,7 +24,6 @@ import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableIndexModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableModel;
 import com.sap.xsk.hdb.ds.parser.XSKDataStructureParser;
 import com.sap.xsk.hdb.ds.transformer.HDBTableDefinitionModelToHDBTableColumnModelTransformer;
-import com.sap.xsk.hdb.ds.transformer.HDBTableDefinitionModelToHDBTableIndexModelTransformer;
 import com.sap.xsk.parser.hdbtable.core.HdbtableLexer;
 import com.sap.xsk.parser.hdbtable.core.HdbtableParser;
 import com.sap.xsk.parser.hdbtable.custom.XSKHDBTABLECoreVisitor;
@@ -32,11 +31,14 @@ import com.sap.xsk.parser.hdbtable.custom.XSKHDBTABLESyntaxErrorListener;
 import com.sap.xsk.parser.hdbtable.exceptions.XSKHDBTableMissingPropertyException;
 import com.sap.xsk.parser.hdbtable.model.XSKHDBTABLEColumnsModel;
 import com.sap.xsk.parser.hdbtable.model.XSKHDBTABLEDefinitionModel;
+import com.sap.xsk.parser.hdbtable.model.XSKHDBTABLEIndexesModel;
 import com.sap.xsk.utils.XSKCommonsConstants;
 import com.sap.xsk.utils.XSKCommonsUtils;
+import com.sap.xsk.utils.XSKConstants;
 import com.sap.xsk.utils.XSKHDBUtils;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.regex.Matcher;
@@ -52,7 +54,6 @@ public class XSKTableParser implements XSKDataStructureParser<XSKDataStructureHD
 
   private static final Logger logger = LoggerFactory.getLogger(XSKTableParser.class);
   private HDBTableDefinitionModelToHDBTableColumnModelTransformer columnModelTransformer = new HDBTableDefinitionModelToHDBTableColumnModelTransformer();
-  private HDBTableDefinitionModelToHDBTableIndexModelTransformer indexModelTransformer = new HDBTableDefinitionModelToHDBTableIndexModelTransformer();
 
   @Override
   public String getType() {
@@ -144,13 +145,48 @@ public class XSKTableParser implements XSKDataStructureParser<XSKDataStructureHD
       }
     });
 
-    if (hdbtableDefinitionModel.getIndexes() != null) {
-      XSKDataStructureHDBTableConstraintUniqueModel uniqueIndex = new XSKDataStructureHDBTableConstraintUniqueModel();
-      XSKDataStructureHDBTableIndexModel tableIndex = new XSKDataStructureHDBTableIndexModel();
+    List<XSKDataStructureHDBTableConstraintUniqueModel> uniqueIndices = new ArrayList<>();
+    List<XSKDataStructureHDBTableIndexModel> indexes = new ArrayList<>();
 
-      dataStructureHDBTableModel.setIndexes(indexModelTransformer.transform(hdbtableDefinitionModel, location, tableIndex));
-      dataStructureHDBTableModel.getConstraints().setUniqueIndices(indexModelTransformer.transform(hdbtableDefinitionModel, location, uniqueIndex));
+    if (hdbtableDefinitionModel.getIndexes() != null) {
+      for (XSKHDBTABLEIndexesModel index : hdbtableDefinitionModel.getIndexes()) {
+        validateIndex(hdbtableDefinitionModel, location, index);
+        if (index.isUnique()) {
+          XSKDataStructureHDBTableConstraintUniqueModel uniqueIndexModel = new XSKDataStructureHDBTableConstraintUniqueModel();
+          if(index.getIndexName() != null){
+            uniqueIndexModel.setName(index.getIndexName());
+          }
+          if (index.getIndexType() != null) {
+            uniqueIndexModel.setIndexType(this.convertIfHanaClassicSyntax(index));
+          }
+          if (index.getOrder() != null) {
+            uniqueIndexModel
+                .setOrder(index.getOrder().equals(XSKConstants.HDBTABLE_INDEX_ORDER_HANA_V1_DSC) ? XSKConstants.HDBTABLE_INDEX_ORDER_HANA_DESC : index.getOrder());
+          }
+          uniqueIndexModel.setColumns(index.getIndexColumns().toArray(String[]::new));
+
+          uniqueIndices.add(uniqueIndexModel);
+        }else{
+          XSKDataStructureHDBTableIndexModel tableIndexModel = new XSKDataStructureHDBTableIndexModel();
+          if(index.getIndexName() != null){
+            tableIndexModel.setIndexName(index.getIndexName());
+          }
+          if (index.getIndexType() != null) {
+            tableIndexModel.setIndexType(convertIfHanaClassicSyntax(index));
+          }
+          tableIndexModel.setUnique(index.isUnique());
+          if (index.getOrder() != null) {
+            tableIndexModel
+                .setOrder(index.getOrder().equals(XSKConstants.HDBTABLE_INDEX_ORDER_HANA_V1_DSC) ? XSKConstants.HDBTABLE_INDEX_ORDER_HANA_DESC : index.getOrder());
+          }
+          tableIndexModel.setIndexColumns(index.getIndexColumns());
+          indexes.add(tableIndexModel);
+        }
       }
+      dataStructureHDBTableModel.getConstraints().setUniqueIndices(uniqueIndices);
+      dataStructureHDBTableModel.setIndexes(indexes);
+    }
+
     return dataStructureHDBTableModel;
   }
 
@@ -161,6 +197,40 @@ public class XSKTableParser implements XSKDataStructureParser<XSKDataStructureHD
         XSKDBContentType.OTHERS);
     dataStructureHDBTableModel.setRawContent(content);
     return dataStructureHDBTableModel;
+  }
+
+  private void validateIndex(XSKHDBTABLEDefinitionModel hdbtableDefinitionModel, String location, XSKHDBTABLEIndexesModel index) {
+    try {
+      index.checkForAllIndexMandatoryFieldsPresence();
+    } catch (Exception e) {
+      XSKCommonsUtils.logCustomErrors(location, XSKCommonsConstants.PARSER_ERROR, "", "", e.getMessage(),
+          XSKCommonsConstants.EXPECTED_FIELDS, XSKCommonsConstants.HDB_TABLE_PARSER, XSKCommonsConstants.MODULE_PARSERS,
+          XSKCommonsConstants.SOURCE_PUBLISH_REQUEST, XSKCommonsConstants.PROGRAM_XSK);
+      throw new XSKHDBTableMissingPropertyException(
+          String.format("Wrong format of table definition: [%s]. [%s]", location, e.getMessage()));
+    }
+
+    index.getIndexColumns().forEach(col -> {
+      List<XSKHDBTABLEColumnsModel> foundMatch = hdbtableDefinitionModel.getColumns().stream().filter(x -> x.getName().equals(col))
+          .collect(Collectors.toList());
+      if (foundMatch.size() != 1) {
+        String errMsg = String.format("%s: the column does not have a definition but is specified as an index", col);
+        XSKCommonsUtils.logCustomErrors(location, XSKCommonsConstants.PARSER_ERROR, "", "", errMsg,
+            "", XSKCommonsConstants.HDB_TABLE_PARSER, XSKCommonsConstants.MODULE_PARSERS,
+            XSKCommonsConstants.SOURCE_PUBLISH_REQUEST, XSKCommonsConstants.PROGRAM_XSK);
+        throw new IllegalStateException(errMsg);
+      }
+    });
+  }
+
+  private String convertIfHanaClassicSyntax(XSKHDBTABLEIndexesModel index) {
+    if (index.getIndexType().equals("CPB_TREE")) {
+      return "CPBTREE";
+    } else if (index.getIndexType().equals("B_TREE")) {
+      return "BTREE";
+    } else {
+      return index.getIndexType();
+    }
   }
 
 }
